ocr/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.15.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
billiard==4.2.1
cattrs==24.1.2
celery==5.5.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
coremltools==8.1
cycler==0.12.1
dill==0.3.8
distro==1.9.0
environs==11.2.1
fastapi==0.115.12
fastapi_cors==0.0.6
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.5.4
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mlflow==3.3.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.0.2
openai==1.98.0
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psycopg2-binary==2.9.10
pyaml==24.12.1
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.9.20
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
tenacity==8.2.3
asyncpg==0.29.0
xxhash==3.5.0
yarl==1.18.3
pytest==8.4.0
pytest-asyncio==0.24.0
aiosqlite==0.20.0
boto3==1.34.30
